feat: store-request create, delete에서 JWT토큰을 이용하도록 변경

+++ src/store-requests/store-requests.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpStatus, NotFoundException, Param, Post, Put, UseGuards } from "@nestjs/common"
+import { Body, Controller, Delete, ForbiddenException, Get, HttpStatus, NotFoundException, Param, Post, Put, Req, UseGuards } from "@nestjs/common"
 import { CreateStoreRequestDTO } from "./DTO/create-store-request.dto"
 import { ReadStoreRequestDTO } from "./DTO/read-store-request.dto"
 import { StoreRequestsService } from "./store-requests.service"
@@ -8,6 +8,7 @@ import { RolesGuard } from "src/common/custom-decorators/custom-role.guard"
 import { AuthGuard } from "@nestjs/passport"
 import { UserRole } from "src/users/entities/user-role.enum"
 import { Roles } from "src/common/custom-decorators/roles.decorator"
+import { AuthenticatedRequest } from "src/auth/interfaces/authenticated-request.interface"
 
 @Controller('api/store-requests')
 @UseGuards(AuthGuard('jwt'), RolesGuard) // JWT인증, roles guard 적용
@@ -19,8 +20,12 @@ export class StoreRequestsController {
     // 가게 신청서 생성
     @Post('/')
     @Roles(UserRole.MANAGER)
-    async createStoreRequest(@Body() createStoreRequestDTO: CreateStoreRequestDTO): Promise<ApiResponseDTO<void>> {
-        await this.storeRequestsService.createStoreRequest(createStoreRequestDTO)
+    async createStoreRequest(
+        @Req() req: AuthenticatedRequest,
+        @Body() createStoreRequestDTO: CreateStoreRequestDTO
+    ): Promise<ApiResponseDTO<void>> {
+        const user_id = req.user.user_id
+        await this.storeRequestsService.createStoreRequest(user_id, createStoreRequestDTO)
 
         return new ApiResponseDTO(true, HttpStatus.CREATED, 'Request Created Successfully')
     }
@@ -31,7 +36,7 @@ export class StoreRequestsController {
     @Roles(UserRole.ADMIN)
     async readAllStoreRequest(): Promise<ApiResponseDTO<ReadStoreRequestDTO[]>> {
         const storeRequests = await this.storeRequestsService.readAllStoreRequest()
-        if (!storeRequests) { 
+        if (!storeRequests) {
             throw new NotFoundException(`Cannot Find Requests`)
         }
 
@@ -45,7 +50,7 @@ export class StoreRequestsController {
     @Roles(UserRole.MANAGER, UserRole.ADMIN)
     async readStoreRequestById(@Param('request_id') request_id: number): Promise<ApiResponseDTO<ReadStoreRequestDTO>> {
         const storeRequest = await this.storeRequestsService.readStoreRequestById(request_id)
-        if (!storeRequest) { 
+        if (!storeRequest) {
             throw new NotFoundException(`Cannot Find Request by Id ${request_id}`)
         }
 
@@ -68,9 +73,23 @@ export class StoreRequestsController {
     // 가게 신청서 삭제
     @Delete('/:request_id')
     @Roles(UserRole.MANAGER, UserRole.ADMIN)
-    async deleteStoreRequest(@Param('request_id') request_id: number) {
-        await this.storeRequestsService.deleteStoreRequest(request_id)
+    async deleteStoreRequest(
+        @Req() req: AuthenticatedRequest,
+        @Param('request_id') request_id: number
+    ): Promise<ApiResponseDTO<void>> {
+        const foundRequest = await this.storeRequestsService.readStoreRequestById(request_id)
 
+        if (req.user.role === UserRole.ADMIN) {
+            await this.storeRequestsService.deleteStoreRequest(request_id)
+            return new ApiResponseDTO(true, HttpStatus.OK, 'Request Deleted Successfully')
+        }
+
+        if (foundRequest.user.user_id !== req.user.user_id) {
+            throw new ForbiddenException('You Can Only Delete Your Own Store Request.')
+
+        }
+
+        await this.storeRequestsService.deleteStoreRequest(request_id)
         return new ApiResponseDTO(true, HttpStatus.OK, 'Request Deleted Successfully')
     }
